refactor: extract each display mode into its own folder
