+++ packages/skott/bin/open-url.ts
